+++ csatar-plugins/knowledgerepository/classes/xlsx/GamesXlsxImport.php
@@ -142,7 +142,12 @@ class GamesXlsxImport implements ToModel, WithHeadingRow, WithValidation, SkipsO
     {
         $validator->after(function ($validator) {
             if (!empty($this->errors) && isset($this->errors[$this->getRowNumber()])) {
-                $validator->errors()->add($this->getRowNumber(), $this->errors[$this->getRowNumber()]);
+                $error = $this->errors[$this->getRowNumber()];
+                if (is_array($error)) {
+                    $error = implode("\n", $error);
+                }
+
+                $validator->errors()->add($this->getRowNumber(), $error);
             }
         });
     }